feat: added directory filtering to only remove .jpg files from tmp/ dir

+++ scripts/real_time_vision.py
@@ -324,15 +324,16 @@ def stream_process(file_dir, nb_frames_sent,dict,lock) :
 
 def cleanup_dir(dir: str):
     for filename in os.listdir(dir):
-        file_path = os.path.join(dir, filename)
-        try:
-            if os.path.isfile(file_path) or os.path.islink(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                shutil.rmtree(file_path)
-            print('Deleted {}'.format(file_path))        
-        except Exception as e:
-            print('Failed to delete %s. Reason: %s' % (file_path, e))
+        if filename.endswith(".jpg"): # only remove jpg files
+            file_path = os.path.join(dir, filename)
+            try:
+                if os.path.isfile(file_path) or os.path.islink(file_path):
+                    os.unlink(file_path)
+                elif os.path.isdir(file_path):
+                    shutil.rmtree(file_path)
+                print('Deleted {}'.format(file_path))        
+            except Exception as e:
+                print('Failed to delete %s. Reason: %s' % (file_path, e))
 
 
 def create_video(image_folder: str, video_name: str):
